EN-865 improve data service object
